+++ BedrockClient/BedrockClient.csproj
@@ -81,9 +81,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Args.cs" />
     <Compile Include="ClientConnector.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ServerInfo.cs" />
     <Compile Include="ServerProcessHelper.cs" />
     <Compile Include="ThreadPayload.cs" />
   </ItemGroup>
